requirements.txt
numpy>=1.24
Pillow>=10.0
boto3>=1.34
gspread>=6.0
google-auth>=2.0
python-dotenv>=1.0
flask
flask-cors
